GetStateDir(): make this a singleton via sync.Once
This function uses the GPUPGRADE_HOME or HOME environment variables, so
we must latch the value on the first call of any process and keep it.
Otherwise, havoc occurs: the state directory might end up being
inconsistent or the program might behave in unexpected ways.

Because the existing initialize_test.go file used GPUPGRADE_HOME twice
with different values, we split those tests into diffferent files.
